telegram-agent-service/requirements.txt
telethon
apscheduler
python-dotenv
requests
pytz
fastapi
uvicorn
aiohttp
